Add zkill, eve_xml, esi modules
